Metadata search funcitons in separate file. + using search keywords instead of full query + removed matchType, relying only on score
